requirements_latex.txt

Flask==2.3.3
python-pptx==0.6.21
reportlab==4.0.4
Werkzeug==2.3.7
